staging: lustre: libcfs: return size_t for libcfs_ioctl_packlen
Change return value to size_t.

Intel-bug-id: https://jira.hpdd.intel.com/browse/LU-6245
Reviewed-on: http://review.whamcloud.com/17643

+++ drivers/staging/lustre/lnet/libcfs/linux/linux-module.c
@@ -41,9 +41,9 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 
 #define LNET_MINOR 240
 
-static inline int libcfs_ioctl_packlen(struct libcfs_ioctl_data *data)
+static inline size_t libcfs_ioctl_packlen(struct libcfs_ioctl_data *data)
 {
-	int len = sizeof(*data);
+	size_t len = sizeof(*data);
 
 	len += cfs_size_round(data->ioc_inllen1);
 	len += cfs_size_round(data->ioc_inllen2);
